Access version page without existing session
If the user doesn't have an existing cookie with an auth token, create a new auth token when they visit the version page. This is so they can get information about the Dynamics system without having to generate a cookie elseewhere in the service.
