+++ src/controllers/root.controller.js
@@ -1,13 +1,7 @@
 'use strict'
 
 const Constants = require('../constants')
-
-// Used for generating a session id which is saved as a cookie
-const uuid4 = require('uuid/v4')
-
 const BaseController = require('./base.controller')
-const ActiveDirectoryAuthService = require('../services/activeDirectoryAuth.service')
-const authService = new ActiveDirectoryAuthService()
 
 module.exports = class RootController extends BaseController {
   static async doGet (request, reply, errors) {
@@ -17,26 +11,7 @@ module.exports = class RootController extends BaseController {
   }
 
   static async doPost (request, reply, errors) {
-    // Generate a session token
-    const token = uuid4()
-
-    // Generate a CRM token
-
-    let authToken
-    try {
-      authToken = await authService.getToken()
-    } catch (err) {
-      console.error(err)
-      return reply.redirect('/error')
-    }
-
-    // TODO: Confirm how session handling will work and where the most
-    // appropriate point is to create and destroy session cookies
-
-    const cookie = {
-      token: token,
-      authToken: authToken
-    }
+    const cookie = await this.generateCookie(reply)
 
     return reply
       .redirect(Constants.Routes.SITE.path)
