+++ src/controllers/version.controller.js
@@ -12,8 +12,14 @@ module.exports = class VersionController extends BaseController {
       const pageContext = BaseController.createPageContext(Constants.Routes.VERSION.pageHeading)
 
       let authToken
+
+      // If a cookie already exists, use that token
       if (request.state[Constants.COOKIE_KEY]) {
         authToken = request.state[Constants.COOKIE_KEY].authToken
+      // Otherwise, create a new one
+      } else {
+        let newCookie = await this.generateCookie(reply)
+        authToken = newCookie.authToken
       }
 
       pageContext.dynamicsSolution = await DynamicsSolution.get(authToken)
@@ -32,6 +38,6 @@ module.exports = class VersionController extends BaseController {
   }
 
   static handler (request, reply, source, errors) {
-    return BaseController.handler(request, reply, errors, VersionController)
+    return BaseController.handler(request, reply, errors, VersionController, false)
   }
 }
